Prefer the group keyboard.
If we're adding a new keyboard group to the seat and don't already have
a seat keyboard, make the newly create group keyboard the seat's keyboard.

This prevens us from bouncing back and forth as new devices are added
to the group, or when new input comes in -- usually from the group
keyboard -- causing Xwayland to reset various key settings such as
autorepeat on/off and rate.